requirements.txt
beautifulsoup4==4.11.1
Flask==3.0.1
Requests==2.31.0
gunicorn==21.2.0
lxml==4.9.2
flask-cors==4.0.0
newspaper3k==0.2.8
nltk==3.8.1
validators==0.22.0
Flask-Limiter==3.5.1
python-dotenv==1.0.1